Break out class definitions
