+++ loups/common_components/__init__.py
@@ -1,0 +1,5 @@
+"""loups/common_components module __init__ file."""
+
+from .shared_classes import Point, Size
+
+__all__ = ["Point", "Size"]
