+++ loups/common_components/shared_classes.py
@@ -1,0 +1,6 @@
+"""Shared classes used accross the package."""
+
+from collections import namedtuple
+
+Point = namedtuple("Point", "x, y")
+Size = namedtuple("Size", "height, width")
